Handle tagging import and refactor
