+++ singlecellmultiomics/universalBamTagger/__init__.py
@@ -1,2 +1,3 @@
 from .universalBamTagger import *
 from .customreads import *
+from .tagging import *
