+++ singlecellmultiomics/universalBamTagger/bamtagmultiome_multi.py
@@ -48,62 +48,12 @@ argparser.add_argument('-blacklist', type=str, help='blacklist (bed file), with 
 argparser.add_argument('-debug_job_bin_bed', type=str, help='Path to write a bed file')
 
 
+def run_tagging(**kwargs):
+    pass
+
 def run_multiome_tagging_cmd(commandline):
     args = argparser.parse_args(commandline)
     run_multiome_tagging(args)
-
-
-
-
-
-def run_tagging_tasks(args):
-    """ Run tagging for one or more tasks
-
-    """
-
-    (alignments_path, temp_dir, timeout_time), arglist = args
-
-    target_file = f"{temp_dir}/{uuid.uuid4()}.bam"
-
-    def timeout_check_function(iteration, mol_iter, reads ):
-        nonlocal time_start
-        nonlocal timeout_time
-        if (datetime.now()-time_start).total_seconds() > timeout_time:
-            raise TimeoutError()
-
-    timeout_tasks = []
-    total_molecules = 0
-    read_groups = dict()
-    with pysam.AlignmentFile(alignments_path) as alignments:
-        with sorted_bam_file(target_file, origin_bam=alignments, mode='wbu',
-            fast_compression=True, read_groups=read_groups) as output:
-            for task  in arglist:
-                # Set a callback function used to check if a problematic region is reached
-                task['molecule_iterator_args']['progress_callback_function'] = timeout_check_function
-
-                try:
-                    statistics = run_tagging_task(alingments, output, read_groups=read_groups, **task)
-                    total_molecules += statistics.get('total_molecules_written',0)
-                except TimeoutError:
-                    timeout_tasks.append( task )
-
-
-    meta = {
-        'timeout_tasks' : timeout_tasks,
-        'total_molecules' : total_molecules,
-    }
-
-    if total_molecules>0:
-        return target_file, meta
-    else:
-        # Clean up ?
-        try:
-            os.remove(target_file)
-            os.remove(f'{target_file}.bai')
-        except Exception as e:
-                pass
-
-        return None, meta
 
 
 def run_multiome_tagging(args):
